refactor(desktop): isolate project tree presentation helpers
